Fix an xpatch so it will work in the GitHub Action.
The problem was that the container used in the GitHub Action
had a newer version of LaTeX, which already has the bugfix that
the patch manually applies. So the tokens that the patch was trying
to replace weren't there, and the patch failed. The failure had an
\errmessage in, so this stopped the whole build. With the
\errmessage removed, the patch fails inconsequentially. But because
the server runs the pre-fixed version, this shouldn't be a problem.