Fix a bug in the Makefile.
I had mistakenly used the sed rule
s|[^ ]*|&/+.tex|g
instead of
s|[^ ]+|&/*.tex|g

The + was there in an effort to ensure that an empty string was
not matched, as in b57d567a571a0dfd08b999f4852a158f570a6c35.
However, it was in the wrong place, attached to the string to be
inserted, not to the string to be searched for. This not only
failed to have the desired effect, but broke the entire
prerequisite. The * that should have been inserted was subsequently
used as a wildcard. But the + that was inserted instead had no
meaning in a wildcard. As such, this prerequisite matched nothing.