Rename the index to "Index of Feats".
The index was originally intended to be a true index. I had
postponed sorting out the entries to some later date, adding only
feats, because it was easy to include all the feats. However,
I have realised that these ought to have their own, dedicated
index *anyway*, and may indeed be the only index necessary.
An index of feats is certainly necessary, even if the books never
go to print, because people might only write down the feat names.
And searching the index might be easier than Ctrl+F, because it
won't find instances of the feat as a prerequisite.
For now, simply leaving this as the only index should serve well.
The links within the book will suffice for most other purposes.