Change the form of the Projection and Medium's Potions.
Rename both feats in the process, and change the structure of
the feats' text heavily..
Make both mixtures snuffs created with a still, instead of spikes
created with a cauldron. The use of a still matches their
predecessor, the Garlic Solution. The use of snuff matches the
immediate effect better, as well as giving more impression of a
*kick*.