Fix a bug in make's dropcap rule.
The rule should only serach for chapter files, so it excludes
files such as core/core.tex. it does this be grepping for
"core.tex" and the like. However, the check for
"creatures-of-the-night.tex" also excluded
"introduction-creatures-of-the-night.tex".
Correct this by adding a / at the start of the grep rule, to match
only the entire file name.