Increase the cost of the contracts feat.
This is in response to noting that contracts can be signed under
duress. As such, it is possible to bind someone to eternal,
unbreakable servitude, forcing them to sign the contract on pain
of death.
It may even be worth limiting the time on contracts and penalties,
to balance this.