Reduce the cost of the squirrel familiar.
Although it gains speed, climbing, and attributes compared to a rat
or hamster, it looses keen smell, which is actually quite useful,
and its skill, Athletics, is less useful than Stealth or
Socialising (Socialising is a prerequisite for the
talking-to-people feat). All this means it should probably cost the
same as a rat or hamster.