Make the first contracts feat cheaper.
This is made necessary by the change in default penalty.
With magical compulsion as the default penalty, the first feat was
very powerul. If nothing else, it was a way to magically compel
someone to do anything, as long as they could be mundanely compelled
to sign a contract. It can still serve a similar purpose, but is
less powerful now, as someone can still disobey the contract if
they're willing to suffer the liar's curse.