Reword the mention of feat costs in the character creation guide.
This is a partial reversion of ab4828240caef4c2dae2982583adefcd2d6c685c.
The addition of the mention of familiar advancement (e46f25895c4ef31a8670c1ce3ce69717ef2595b2)
to the same paragraph made the wording awkward, interjecting a disjoint idea
in the middle of the discussion of feat costs. This hopefully helps.