Limit damage tests from strength stimulant to once per turn.
It seems unfair that you'd get two damage tests if you took the
dash action, and then the GM called for a test to jump a chasm.
Limiting it once per round, instead of once per turn, also makes
it less punishing if an enemy forces an opposed test. This also
means that I don't have to specify that damage tests as part of
the attack action don't cause damage tests from the potion; this
occurs implicitly as the attack roll will already have caused a
damage test from the potion.