Remove the poison resistance feat.
While every feat is reasonably a property of the mind, this feat
seemed to be a property of the body, and thus was making writing
the rules. I could have refluffed it to be a more mental thing,
perhaps only working while you're conscious, but then it seems
like it shouldn't even be part of Brewing as a discipline, but
perhaps more part of Willing. So I've just removed it for now.
I might add something similar later.